+++ addon/components/variable-plugin/insert-variable-card.ts
@@ -10,6 +10,10 @@ import {
 import { CodeList } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/utils/fetch-data';
 import { findParentNodeOfType } from '@curvenote/prosemirror-utils';
 import { NodeSelection } from '@lblod/ember-rdfa-editor';
+import { service } from '@ember/service';
+import IntlService from 'ember-intl/services/intl';
+import { isNumber } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/strings';
+
 type Args = {
   controller: SayController;
   options: {
@@ -19,6 +23,28 @@ type Args = {
   };
 };
 
+type minMaxObj = { minimumValue?: number; maximumValue?: number };
+class ExtraAttributes {
+  @tracked minimumValue = '';
+  @tracked maximumValue = '';
+
+  asObject(): minMaxObj {
+    const obj: minMaxObj = {};
+    if (isNumber(this.minimumValue)) {
+      obj.minimumValue = Number(this.minimumValue);
+    }
+    if (isNumber(this.maximumValue)) {
+      obj.maximumValue = Number(this.maximumValue);
+    }
+    return obj;
+  }
+
+  reset() {
+    this.minimumValue = '';
+    this.maximumValue = '';
+  }
+}
+
 export default class EditorPluginsInsertCodelistCardComponent extends Component<Args> {
   @tracked variablesArray: VariableType[];
   @tracked selectedVariable?: VariableType;
@@ -26,9 +52,11 @@ export default class EditorPluginsInsertCodelistCardComponent extends Component<
   @tracked selectedSubtype?: CodeList;
   @tracked subtypes?: CodeList[];
   @tracked variableLabel?: string;
-  @tracked extraAttributes: Record<string, unknown> = {};
+  @tracked extraAttributes = new ExtraAttributes();
   publisher: string;
   endpoint: string;
+
+  @service declare intl: IntlService;
 
   constructor(parent: unknown, args: Args) {
     super(parent, args);
@@ -66,6 +94,20 @@ export default class EditorPluginsInsertCodelistCardComponent extends Component<
     return this.args.controller;
   }
 
+  get numberVariableError() {
+    const minVal = this.extraAttributes.minimumValue;
+    const maxVal = this.extraAttributes.maximumValue;
+    if (
+      isNumber(minVal) &&
+      isNumber(maxVal) &&
+      Number(minVal) > Number(maxVal)
+    ) {
+      return this.intl.t('variable.number.error-min-bigger-than-max');
+    }
+
+    return '';
+  }
+
   @action
   updateVariableLabel(event: InputEvent) {
     this.variableLabel = (event.target as HTMLInputElement).value;
@@ -73,7 +115,7 @@ export default class EditorPluginsInsertCodelistCardComponent extends Component<
 
   @action
   insert() {
-    if (!this.selectedVariable) {
+    if (!this.selectedVariable || this.numberVariableError) {
       return;
     }
 
@@ -83,12 +125,12 @@ export default class EditorPluginsInsertCodelistCardComponent extends Component<
       label: this.variableLabel !== '' ? this.variableLabel : undefined,
       attributes: {
         source: this.endpoint,
-        ...this.extraAttributes,
+        ...this.extraAttributes.asObject(),
       },
     });
 
     this.variableLabel = '';
-    this.extraAttributes = {};
+    this.extraAttributes.reset();
 
     this.controller.withTransaction(
       (tr) => {
@@ -124,7 +166,7 @@ export default class EditorPluginsInsertCodelistCardComponent extends Component<
   @action
   updateSubtype(subtype: CodeList) {
     this.selectedSubtype = subtype;
-    this.extraAttributes = {};
+    this.extraAttributes.reset();
   }
 
   get type() {